+++ config/settings.php
@@ -72,6 +72,7 @@ return [
                 'user'     => '',
                 'password' => '',
                 'charset'  => 'utf8',
+                //'unix_socket'   => '/Applications/MAMP/tmp/mysql/mysql.sock',
             ],
         ],